Use Clion to optimize codes

+++ src/GUI/SoundManager.h
@@ -8,7 +8,7 @@ public:
   SoundManager();
   ~SoundManager();
 
-  void Play(const std::string &l_soundName, int l_volume = 100, bool l_loop = false);
+  void Play(const std::string &l_soundName, float l_volume = 100, bool l_loop = false);
 
 private:
   std::unordered_map<std::string, sf::SoundBuffer> m_buffers;
